+++ NEWS.md
@@ -1,3 +1,12 @@
+# twinetverse 0.0.2
+
+Add 2 functions to prepare data from `graphTweets` for `sigmajs` use.
+
+* `edges2sg`
+* `nodes2sg`
+
+Imports `%<-%` from the zeallot package to unpack network from `graphTweets`.
+
 # twinetverse 0.0.1
 
 Initial version